:gear: Déplace l'injection des traductions vers rules
