+++ source/engine/rules.js
@@ -1,6 +1,8 @@
 // Séparation artificielle, temporaire, entre ces deux types de règles
 import rawRules from 'Règles/base.yaml'
+import translations from 'Règles/externalized.yaml'
 import {
+	assoc,
 	has,
 	pipe,
 	toPairs,
@@ -113,11 +115,6 @@ export let collectDefaults = pipe(
 	fromPairs
 )
 
-// On enrichit la base de règles avec des propriétés dérivées de celles du YAML
-export let rules = rawRules.map(rule =>
-	enrichRule(rule, { taux_versement_transport })
-)
-
 /****************************************
  Méthodes de recherche d'une règle */
 
@@ -168,3 +165,31 @@ export let formatInputs = (flatRules, formValueSelector) => state => name => {
 
 	return value && pre(value)
 }
+
+/* Traduction */
+
+export let translateAll = (translations, flatRules) => {
+	let translationsOf = rule =>
+			translations[rule.dottedName],
+		translateProp = (lang, translation) => (rule, prop) => {
+			let propTrans = translation[prop+"."+lang]
+			return propTrans ?
+				assoc(prop, propTrans, rule) :
+				rule
+		},
+		translateRule = (lang, translations, props) => rule => {
+			let ruleTrans = translationsOf(rule)
+			return ruleTrans ?
+					reduce(translateProp(lang, ruleTrans), rule, props) :
+					rule
+		}
+
+	let targets = ["titre", "description", "question", "sous-question"]
+
+	return map(translateRule("en", translations, targets), flatRules)
+}
+
+// On enrichit la base de règles avec des propriétés dérivées de celles du YAML
+export let rules = translateAll(translations, rawRules.map(rule =>
+	enrichRule(rule, { taux_versement_transport })
+))
