+++ test/rules.test.js
@@ -1,8 +1,9 @@
-import R from 'ramda'
+import { map } from 'ramda'
 import { expect } from 'chai'
 import {
 	rules,
 	enrichRule,
+	translateAll,
 	findVariantsAndRecords
 } from '../source/engine/rules'
 
@@ -60,5 +61,33 @@ describe('rule checks', function() {
 			)
 		)
 		expect(rulesNeedingDefault).to.be.empty
+	})
+})
+
+describe('translateAll', function() {
+	it('should translate flat rules', function() {
+		let rules = [{
+			"espace":"foo",
+			"nom":"bar",
+			"titre":"Titre",
+			"description":"Description",
+			"question":"Question",
+			"sous-question":"Sous Question",
+		}]
+		let translations = {
+			"foo . bar":{
+				"titre.en":"TITRE",
+				"description.en":"DESC",
+				"question.en":"QUEST",
+				"sous-question.en":"SOUSQ",
+			}
+		}
+
+		let result = translateAll(translations, map(enrichRule,rules))
+
+		expect(result[0]).to.have.property("titre","TITRE")
+		expect(result[0]).to.have.property("description","DESC")
+		expect(result[0]).to.have.property("question","QUEST")
+		expect(result[0]).to.have.property("sous-question","SOUSQ")
 	})
 })
